% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accumulateMeasurements}
\alias{accumulateMeasurements}
\title{Accumulate measurements}
\usage{
accumulateMeasurements(dataset, config)
}
\arguments{
\item{dataset}{a data frame with corrected and calibrated measurement data of
a specific data set.}

\item{config}{A named list of configuration parameters (e.g. as read from the
\code{config.yaml} file) containing at least the component
\code{average_over_inj}.}
}
\value{
A data frame with nine columns with the injection-averaged values of
  d18O, dH and d-excess together with their standard deviations as well as
  the respective \code{Identifier 1}, \code{Identifier 2} and \code{block}
  specifiers.
}
\description{
Calculate the average across a specified number of injections of the d18O,
dH and d-excess values for each measured sample of a dataset and calculate
the standard deviation of the means.
}
\details{
The functions uses the config parameter \code{average_over_inj}, which
specifies the injections to average. If it is -1 or 'all', all injections are
used, if it is set to a single integer `n`, the last `n` injections are used,
else if it is set to a range `n1:n2`, injections `n1:n2` are used.
}
