% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addColumnDExcess}
\alias{addColumnDExcess}
\title{Calculate d-excess}
\usage{
addColumnDExcess(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}
}
\value{
The input \code{dataset} supplemented by the column \code{dExcess}.
}
\description{
Calculate the second-order parameter d-excess from the d18O and dD values of
a given data set according to \code{d-excess = dD - 8 * d18O}.
}
