% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addColumnSecondsSinceStart}
\alias{addColumnSecondsSinceStart}
\title{Calculate seconds since start}
\usage{
addColumnSecondsSinceStart(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}
}
\value{
The input \code{dataset} supplemented by the column
\code{SecondsSinceStart} which gives the seconds elapsed since the start of
the measurement.
}
\description{
Calculate the seconds elapsed since the start of the measurement.
}
\details{
The input data frame is expected to contain the column
\code{Time Code} with values as character vectors of the format
'yyyy/mm/ddhh:mm:ss' (e.g. '2019/11/2510:00:00').
}
