% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assignVialsToGroups}
\alias{assignVialsToGroups}
\title{Assign vial groups}
\usage{
assignVialsToGroups(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set;
needs to contain at least the columns \code{Line} and \code{Identifier 1}.}
}
\value{
The input \code{dataset} appended by the column \code{vial_group}.
}
\description{
This function adds the additional column \code{vial_group} to the input
data frame counting the occurrence of groups of consecutive vials of the
same standard or sample across the measurement.
}
