% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{buildFirstSection}
\alias{buildFirstSection}
\title{Build section "AVERAGE OVER ALL FILES"}
\usage{
buildFirstSection(processedData)
}
\arguments{
\item{processedData}{a list as output by \code{processData()}.}
}
\description{
Build section "AVERAGE OVER ALL FILES"
}
