% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{buildThirdSection}
\alias{buildThirdSection}
\title{Build section "INTER STANDARD BIAS TO LITERATURE VALUES FOR EACH FILE"}
\usage{
buildThirdSection(processedData, config)
}
\arguments{
\item{processedData}{a list as output by \code{processData()}.}

\item{config}{a named list with the contents of the config file.}
}
\description{
Build section "INTER STANDARD BIAS TO LITERATURE VALUES FOR EACH FILE"
}
