% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory-correction.R
\name{calculateMemoryCoefficients}
\alias{calculateMemoryCoefficients}
\title{Calculate memory coefficients}
\usage{
calculateMemoryCoefficients(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block} and
\code{vial_group} which are not included in the raw Picarro output.}
}
\value{
A data frame with rows corresponding to injection numbers and columns
  to standards and isotopic species; specifically, for each unique vial block
  of standards, memory coefficients for d18O and dD are returned. The
  mean memory coefficients are returned in the columns \code{memoryCoeffD18O}
  (d18O) and \code{memoryCoeffDD} (dD). See also the package vignette for
  more details.
}
\description{
Calculate memory coefficients using the data of standards measured in the
first standard block at the beginning of the measurement sequence.
}
\details{
Memory coefficients are estimated based on the formula in
\code{\link{formulaMemCoeff}} for the (n - 1) of the n standards in the first
standard block (per definition, no coefficients can be calulated for the very
first standard in the block), and the average across these (n - 1) individual
sets of memory coefficients is calculated. The maximum number of mean memory
coefficients is given by the maximum number of injections of the used
standards.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{assignVialsToGroups}}
}
