% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calculatePoooledStdDev.R
\name{calculatePoooledStdDev}
\alias{calculatePoooledStdDev}
\title{calculatePoooledStdDev}
\usage{
calculatePoooledStdDev(datasets)
}
\arguments{
\item{datasets}{A named list of dataframes with isotope measurement data.}
}
\value{
A list. The list elements are named like the input list "datasets". 
        Each element of the list is a list with the two named elements
        "d18O" and "dD" (the pooled standard deviation for d18O and dD).
}
\description{
Calculate the pooled standard deviation for d18O and dD. The pooled standard
deviation is calculated using the following formulas:
}
\details{
x = (n_1 - 1) \* sd_1^2 + ... + (n_k - 1) \* sd_k^2
    y = n_1 + ... + n_k - k
    sdPooled = sqrt( x / y )
}
