% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_DataCorrCalib.R
\name{CalibratePicarroData}
\alias{CalibratePicarroData}
\title{Correct and calibrate raw data}
\usage{
CalibratePicarroData(folder.iso_data, flag.MemCorr, flag.Calibration)
}
\arguments{
\item{folder.iso_data}{a list with two elements: (1) the isotope data as the
result from the call to \code{\link{wrapper.ReadDataFiles}}; (2) the meta
information on the measurement sequence as the result from the call to
\code{\link{get.input_folder}}.}

\item{flag.MemCorr}{integer; flag to signal if a memory correction shall be
applied (1 = YES, 0 = NO)}

\item{flag.Calibration}{integer; flag to signal which calibration scheme
shall be applied (1 = linear drift correction + single three-point
calibration at the beginning of the measurement sequence; 2 = double
three-point calibration at the beginning and at the end of the measurement
sequence).}
}
\value{
a list structure with the corrected and calibrated isotope data
together with statistical data on the quality of the approach.
}
\description{
Main function to correct and calibrate the raw Picarro data.
}
\author{
Thomas Münch
}
