% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory-correction.R
\name{correctForMemoryEffect}
\alias{correctForMemoryEffect}
\title{Correct for memory effect}
\usage{
correctForMemoryEffect(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block} and
\code{vial_group} which are not included in the raw Picarro output.}
}
\value{
A named list with two elements:
  \describe{
  \item{\code{datasetMemoryCorrected}:}{the input \code{dataset} with the
    d18O and dD columns corrected for memory effect.}
  \item{\code{memoryCoefficients}:}{a data frame of estimated memory
    coefficients (see \code{\link{calculateMemoryCoefficients}} for
    details).}
}
}
\description{
Correct a given measurement data set for the memory effect by applying
memory coefficients estimated based on analysing the standards in the first
standard block.
}
\seealso{
\code{\link{calculateMemoryCoefficients}},
  \code{\link{groupStandardsInBlocks}},
  \code{\link{assignVialsToGroups}}
}
