% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{createOutputDirectory}
\alias{createOutputDirectory}
\title{Create output directory}
\usage{
createOutputDirectory(folder)
}
\arguments{
\item{folder}{character string with the path to the folder which is
to be created. The folder may already exist in which case nothing happens.}
}
\value{
the input \code{folder}.
}
\description{
Create the directory in which the programme output data is saved.
}
