% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{CSV.Output}
\alias{CSV.Output}
\title{csv output}
\usage{
CSV.Output(file_data, out_dir, flag.CSV_Output)
}
\arguments{
\item{file_data}{a data frame with the isotope data to write.}

\item{out_dir}{character string with the path to the output directory.}

\item{flag.CSV_Output}{integer; flag to signal which output scheme
to use (1 = only output the sample data; 2 = output sample as well
as standard data in the original order of the read files).}
}
\description{
Wrapper function to produce the csv file output of the corrected and
calibrated isotope data.
}
\author{
Thomas Münch
}
