% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doAccumulate}
\alias{doAccumulate}
\title{Average over injections}
\usage{
doAccumulate(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}
}
\value{
A data frame with nine columns with the injection-averaged values of
  d18O, dH and d-excess together with their standard deviations as well as
  the respective \code{Identifier 1}, \code{Identifier 2} and \code{block}
  specifiers.
}
\description{
This function takes an input data frame and returns the average across all
included injections for each sample and the respective standard deviations.
}
