% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftCorrection.R
\name{DriftCorrection}
\alias{DriftCorrection}
\title{Apply drift correction}
\usage{
DriftCorrection(block_data, params)
}
\arguments{
\item{block_data}{the standard data of the isotopic species in the given
measurement block.}

\item{params}{needed parameters; see the calls to this function in
\code{LinearDriftCorrection.standard_data()} for the details.}
}
\value{
the drift-corrected block data.
}
\description{
Apply a drift correction on the standard data for a given isotopic species
and for a certain measurement block within the measurement sequence.
}
\author{
Thomas Münch
}
