% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{ExtractBlockVariable}
\alias{ExtractBlockVariable}
\title{Extract block variable}
\usage{
ExtractBlockVariable(x, std_data, block.no, var)
}
\arguments{
\item{x}{character string with the name of the standard for which to extract
\code{var} (I think...).}

\item{std_data}{data frame with the standard data from a certain measurement
file.}

\item{block.no}{the number of the block within the measurement sequence.}

\item{var}{the variable to extract.}
}
\value{
the extracted variable.
}
\description{
Extract a certain variable for a given standard from the standard data of a
given standard block of the measurement sequence (*sigh*).
}
\author{
Thomas Münch
}
