% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{gatherQualityControlInfo}
\alias{gatherQualityControlInfo}
\title{Quality control information across processing run}
\usage{
gatherQualityControlInfo(datasets)
}
\arguments{
\item{datasets}{the processed measurement data as output by
\code{\link{processFiles}} (or by \code{\link{processData}} directly).}
}
\value{
A named list with seven elements:
\describe{
  \item{\code{rmsdQualityControl}:}{a tibble for the root-mean-square
  deviation (rmsd) of the quality control standard(s) for each measurement
  file, listing the file name, the name(s) of the quality control
  standard(s), and the rmsd values for the oxygen and hydrogen isotope
  values.}
  \item{\code{rmsdAllStandards}:}{a tibble for the root-mean-square
  deviation (rmsd) of all measured standards for each measurement file,
  listing the file name and the rmsd values for the oxygen and hydrogen
  isotope values.}
  \item{\code{pooledSD}:}{a tibble for the pooled standard deviation across
  the samples measured in each measurement file, listing the file name and
  the pooled SD values for the oxygen and hydrogen isotope values.}
  \item{\code{deviationsFromTrue}:}{a list of tibbles, where each tibble
  lists the deviations of all measured standards from their true values for a
  processed measurement file.}
  \item{\code{memoryCoefficients}:}{a tibble of the mean and the SD of the
  average memory coefficients depending on the injection number per each
  processed measurement file and of the mean and the SD of the overall
  average memory coefficients; or \code{NULL} if no memory correction was
  applied to the measurement data.}
  \item{\code{calibrationParameter}:}{a tibble with information on the
  applied calibration regression, including calibration slope and intercept,
  p-values, r-squared value, and the root mean square deviation of the
  calibration residuals. The information are included for each estimated
  calibration, i.e., for each processed measurement file, and per file for
  each used calibration block and each isotope species.}
  \item{\code{driftParameter}:}{a tibble with information on an applied
  simple drift correction, or \code{NULL} if no such correction was applied
  to the measurement data. The drift correction information include the drift
  correction slope, its p-value, the r-squared value of the linear
  regression, and the root mean square deviation of the regression
  residuals. The information are included for each estimated drift
  correction, i.e., for each processed measurement file, and per file for
  each isotope species and each used drift monitoring standard as well as the
  average across those standards.}
}
}
\description{
This is an auxiliary function to gather the various quality control
information determined for each processed measurement file across all such
files from a given processing run.
}
