% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{get.input_folder}
\alias{get.input_folder}
\title{Read in meta information file.}
\usage{
get.input_folder(dir)
}
\arguments{
\item{dir}{the directory path as a character string where the meta
information file is located.}
}
\value{
a list with the information read in from \code{INPUT.txt}.
}
\description{
Read in the meta information file \code{INPUT.txt} containing necessary
information on the measurement sequence and the scheme of how to correct the
data.
}
\author{
Thomas Münch
}
