% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{getCalibInterceptAndSlope}
\alias{getCalibInterceptAndSlope}
\title{Get calibration parameters}
\usage{
getCalibInterceptAndSlope(dataset, config, useBlock)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{o18_True} and \code{H2_True} which are not
included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}
}
\value{
A named list with elements \code{d18O} and \code{dD} where each
element is again a list with two elements:
\describe{
\item{\code{intercept}:}{numeric; the estimated intercept of the
  calibration.}
\item{\code{slope}:}{numeric; the estimated slope of the calibration.}
}
}
\description{
Obtain calibration slope and intercept values as estimated from standards in
a particular standard block of the measurement sequence from regressing the
expected standard values against their measured values.
}
\details{
If memory correction was applied to the data, all injections are used from
the standard data, else only the last three injections. If a two-point
calibration is switched on, only the two standards with the lowest and
highest isotopic values are used, regardless of how many calibration
standards are actually set in the dataset.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}}
}
