% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getCalibrationSlopes}
\alias{getCalibrationSlopes}
\title{Temporal change of calibration parameters}
\usage{
getCalibrationSlopes(params)
}
\arguments{
\item{params}{a tibble with the calibration parameters for d18O and dD (see
\code{\link{getCalibration}} for details on the tibble structure) as
estimated from the first and the final standard block in the measurement
sequence.}
}
\value{
A named list with elements \code{d18O} and \code{dD} where each
element is again a list with two elements:
\describe{
\item{\code{alpha}:}{numeric; the slope of the linear change in the
  calibration intercept across the measurement sequence.}
\item{\code{beta}:}{numeric; the slope of the linear change in the
  calibration slope across the measurement sequence.}
}
}
\description{
Estimate the change in calibration parameters between the beginning and the
end of the measurement sequence based on a simple linear regression of the
calibration parameters against the elapsed measurement time.
}
