% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCalibTimes}
\alias{getCalibTimes}
\title{Average measurement time of blocks}
\usage{
getCalibTimes(dataset, useBlocks)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional column \code{block} which is not included
in the raw Picarro output.}

\item{useBlocks}{an integer vector specifying the numbers of the standard
blocks for which the average time shall be calculated.}
}
\value{
A numeric vector of the same length as \code{useBlocks} with the
  average measurement time elapsed since start of the measurement for the
  respective blocks.
}
\description{
This function calculates the average measurement time that has elapsed for
the specified standard blocks since the start of the measurement sequence.
}
\seealso{
\code{\link{groupStandardsInBlocks}}
}
