% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftCorrection.R
\name{GetDriftCorrSlope}
\alias{GetDriftCorrSlope}
\title{Drift correction slopes}
\usage{
GetDriftCorrSlope(ix.std, std_data, meta.info_std, flag.iso_type)
}
\arguments{
\item{ix.std}{number of the standard for which to estimate the drift
correction slope.}

\item{std_data}{a data frame with only the data of the standard samples for
a specific measurement file.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}

\item{flag.iso_type}{character string of the isotopic species for which to
estimate the drift correction slope; either "delta.O18" or "delta.H2".}
}
\value{
the estimated drift correction slope.
}
\description{
Estimate a linear drift correction slope from standard data.
}
\author{
Thomas Münch
}
