% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetLastN}
\alias{GetLastN}
\title{Last N elements}
\usage{
GetLastN(x, params)
}
\arguments{
\item{x}{a numeric vector.}

\item{params}{a list containing an element \code{N} giving the number of the
last \code{N} elements to extract.}
}
\value{
numeric vector of length \code{N}.
}
\description{
Return the last N elements of a vector.
}
\author{
Thomas Münch
}
