% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetMeanFirstN}
\alias{GetMeanFirstN}
\title{Mean of first N elements}
\usage{
GetMeanFirstN(x, params)
}
\arguments{
\item{x}{a numeric vector.}

\item{params}{a list containing an element \code{N} giving the number of the
first \code{N} elements to average.}
}
\value{
numeric value with the average of the \code{N} elements.
}
\description{
Return the mean of the first N elements of a vector.
}
\author{
Thomas Münch
}
