% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetMeanLastN}
\alias{GetMeanLastN}
\title{Mean of last N elements}
\usage{
GetMeanLastN(x, params)
}
\arguments{
\item{x}{a numeric vector.}

\item{params}{a list containing an element \code{N} giving the number of the
last \code{N} elements to average.}
}
\value{
numeric value with the average of the \code{N} elements.
}
\description{
Return the mean of the last N elements of a vector.
}
\author{
Thomas Münch
}
