% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIn.R
\name{GetStdBlocks}
\alias{GetStdBlocks}
\title{Split standard data into measurement blocks}
\usage{
GetStdBlocks(ix.std, params)
}
\arguments{
\item{ix.std}{number of the isotope standard to extract.}

\item{params}{list structure with needed parameters; see the code.}
}
\value{
a list structure with \code{n} elements where \code{n} are the
number of different measurement blocks; each element is a matrix containing
the isotope data for standard number \code{ix.std}.
}
\description{
Split the isotope data of the read in standrad sample measurements into the
respective measurement blocks, i.e. the standard measurements at the
beginning, the end and possible in between of the measurement sequence.
}
\author{
Thomas Münch
}
