% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetSubArr}
\alias{GetSubArr}
\title{See PassVector}
\usage{
GetSubArr(x, FUN, params)
}
\arguments{
\item{x}{numeric vector.}

\item{FUN}{function to apply to \code{x}.}

\item{params}{additional parameters passed on to \code{FUN}.}
}
\description{
See \code{\link{PassVector}}!?
}
\author{
Thomas Münch
}
