% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetSubList}
\alias{GetSubList}
\title{Pass list to lapply}
\usage{
GetSubList(x, wrapperFUN, FUN, params)
}
\arguments{
\item{x}{a list.}

\item{wrapperFUN}{function to use in call to \code{lapply}.}

\item{FUN}{function argument passed on to \code{wrapperFUN}.}

\item{params}{additional parameters passed on to \code{FUN}.}
}
\value{
a list of the results of \code{wrapperFUN} applied on \code{x}.
}
\description{
Pass list to lapply with specified functions.
}
\author{
Thomas Münch
}
