% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{GetSubVec}
\alias{GetSubVec}
\title{Pass object to apply}
\usage{
GetSubVec(x, FUN, params)
}
\arguments{
\item{x}{an object that can be passed to \code{apply}.}

\item{FUN}{function to use in call to \code{apply}.}

\item{params}{additional parameters passed on to \code{FUN}.}
}
\value{
a object of the same type as \code{x} of the results of \code{FUN}
applied on the columns of \code{x}.
}
\description{
Pass an object to apply with specified function.
}
\author{
Thomas Münch
}
