% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getTrainingData}
\alias{getTrainingData}
\title{Get calibration training data}
\usage{
getTrainingData(dataset, config, useBlock)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{o18_True} and \code{H2_True} which are not
included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}

\item{useBlock}{a single integer giving the number of the standard block
which contains the data of the requested standards.}
}
\value{
The input \code{dataset} filtered by the requested standard block
  (\code{useBlock}) and by the standards set as calibration standards in
  \code{config}.
}
\description{
Obtain the standard data to be used for a regression of expected against
measured values in order to estimate calibration slope and intercept.
}
\details{
If memory correction was applied to the data, all injections are used from
the standard data, else only the last three injections. If a two-point
calibration is switched on, only the two standards with the lowest and
highest isotopic values are used, regardless of how many calibration
standards are actually set in the dataset.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}}
}
