% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{groupStandardsInBlocks}
\alias{groupStandardsInBlocks}
\title{Group standards in blocks}
\usage{
groupStandardsInBlocks(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{config}{A named list which needs to contain at least the component
\code{standards} which is expected to be a list containing at least the
component \code{name} giving the name for each used standard as specified
by the \code{Identifier 1} column of the data set.}
}
\value{
The input \code{dataset} supplemented by the column \code{block}
  which counts the number of standard blocks across the measurement. For
  normal samples, this value is set to \code{NA}.
}
\description{
Determine for each standard injection which standard block it belongs to.
}
