% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DriftCorrection.R
\name{LinearDriftCorrection.standard_data}
\alias{LinearDriftCorrection.standard_data}
\title{Drift correction}
\usage{
LinearDriftCorrection.standard_data(std_data, meta.info_std)
}
\arguments{
\item{std_data}{a data frame with only the data of the standard samples for
a specific measurement file.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}
}
\value{
a list element containing \code{std_data} with the drift-corrected
standard data as well as information on the drift correction coefficients.
}
\description{
Obtain drift correction coefficients and apply them on the standard data of
a given measurement file.
}
\author{
Thomas Münch
}
