% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemoryCorrection.R
\name{MemCorrData}
\alias{MemCorrData}
\title{Apply memory correction}
\usage{
MemCorrData(x, file_data, mem.coeff_O18, mem.coeff_H2)
}
\arguments{
\item{x}{a four-element numeric vector with the line number, injection
number, d18O value and d2H value of a specific injection from the
measurement sequence.}

\item{file_data}{data frame of raw isotope data of a specific measurement
file.}

\item{mem.coeff_O18}{numeric vector of memory coefficients for oxygen
isotope data (d18O).}

\item{mem.coeff_H2}{numeric vector of memory coefficients for deuterium
isotope data (d2H).}
}
\value{
a four-element numeric vector as \code{x} but with the
memory-corrected isotope data.
}
\description{
Apply the memory correctino scheme of van Geldern and Barth (2012) on a
given measurement (i.e. injection).
}
\author{
Thomas Münch
}
