% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemoryCorrection.R
\name{MemoryCorrection}
\alias{MemoryCorrection}
\title{Perform memory correction}
\usage{
MemoryCorrection(file_data, std_data, sample_data, meta.info_std)
}
\arguments{
\item{file_data}{data frame of raw isotope data of a specific measurement
file.}

\item{std_data}{subset of \code{file_data} with only the data of the
standard samples.}

\item{sample_data}{subset of \code{file_data} with only the data of the
measurement samples.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}
}
\value{
a list with two data frames with the memory-corrected standard and
sample data.
}
\description{
Function to obtain memory coefficients and to apply them to memory-correct
the sample and standard isotope data.
}
\author{
Thomas Münch
}
