% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeInjectionNumbers}
\alias{normalizeInjectionNumbers}
\title{Re-calculate injection numbers}
\usage{
normalizeInjectionNumbers(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}
}
\value{
The input \code{dataset} with the column of injection numbers
  re-calculated accounting for possible consecutive vials of the same probe.
}
\description{
Re-calculate the injection numbers to account for probes being measured
from two or more consecutive vials.
}
