% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{PassVector}
\alias{PassVector}
\title{Pass vector}
\usage{
PassVector(x, FUN, params)
}
\arguments{
\item{x}{numeric vector.}

\item{FUN}{function to apply to \code{x}.}

\item{params}{additional parameters passed on to \code{FUN}.}
}
\value{
the result of \code{FUN} applied on \code{x}.
}
\description{
Pass vector to function.
}
\author{
Thomas Münch
}
