% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccr-main.R
\name{piccr}
\alias{piccr}
\title{"piccr" main routine}
\usage{
piccr(dir = getwd())
}
\arguments{
\item{dir}{the directory path as a character string where the raw Picarro
data files and the meta information file are located.}
}
\value{
\code{0} returned invisibly for a successful programme run.
}
\description{
Correct and calibrate raw Picarro CRDS data.
}
\author{
Thomas Münch
}
