% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{printQualityControl}
\alias{printQualityControl}
\title{Print quality control information}
\usage{
printQualityControl(
  datasets,
  printDeviations = FALSE,
  n = 3,
  printMemoryCoefficients = TRUE,
  whichMemoryCoefficients = "mean",
  printCalibrationParameter = FALSE,
  printDriftParameter = FALSE
)
}
\arguments{
\item{datasets}{the processed measurement data as output by
\code{\link{processFiles}} (or by \code{\link{processData}} directly).}

\item{printDeviations}{logical to control whether the specific deviations
from the true value of each measured standard are printed for the processed
measurement datasets (defaults to \code{FALSE}).}

\item{n}{integer to control the number of printed datasets if
\code{printDeviations = TRUE}; default is to print the first three
processed datasets, set to \code{NA} to print all of them.}

\item{printMemoryCoefficients}{logical to control whether the estimated
memory coefficients shall be printed; defaults to \code{TRUE}, and only
prints when a memory correction was actually applied.}

\item{whichMemoryCoefficients}{character string to signal which part of the
estimated memory coefficients are printed if \code{printMemoryCoefficients
= TRUE}. Possible options are "mean" (the default) to print only the values
for the overall average memory coefficients or "all" to also print the
values for the average memory coefficients per each processed measurement
data set.}

\item{printCalibrationParameter}{logical to control whether information on
the applied calibration regression shall be printed; defaults to
\code{FALSE}.}

\item{printDriftParameter}{logical to control whether information on
an applied drift correction shall be printed; defaults to \code{FALSE},
and only prints when a drift correction was actually applied.}
}
\value{
The input \code{datasets} are returned invisibly.
}
\description{
Print a summary of the quality control information for a piccr processing run
of several measurement files.
}
\details{
Per default, the function prints the following three sections:
\itemize{
  \item the first section gives the overall values for the entire processing
        run of the root-mean-square deviation (rmsd) of the quality control
        standards, the rmsd of all used standards, and the mean value of the
        pooled standard deviation per sample;
  \item the second section contains tables with the above values for each
        processed measurement file.
  \item the third section contains a table of the overall memory
        coefficients depending on the injection number averaged across the
        processing run together with their estimated standard
        deviations. This is only printed if a memory correction was actually
        applied to the measurement data.
}
Additional information can be switched on via the respective function
parameters, such as the deviations of all measured standards from their true
values, information on the calibration regressions, and information on the
drift correction, if applicable.
}
\examples{
# Most conveniently, this function is combined with the processing function
# in a pipe:
\dontrun{
  library(magrittr)
  library(piccr)

  processFiles("my_config_file.yaml") \%>\% printQualityControl()
}
}
