% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{printRunInfo}
\alias{printRunInfo}
\title{Print piccr run information}
\usage{
printRunInfo(configFile)
}
\arguments{
\item{configFile}{the file name of the YAML configuration file which was
used for the processing.}
}
\description{
Auxiliary function to build the top section of the quality control summary
file printing information on the piccr version which was used for the
processing run, the name of the used YAML configuration file, and the
processing date.
}
\seealso{
\code{\link{outputSummaryFile}}
}
