% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_accumulateMeasurementsForEachSample.R
\name{processDataForOutput}
\alias{processDataForOutput}
\title{Process data for output}
\usage{
processDataForOutput(datasets, config)
}
\arguments{
\item{datasets}{A named list of data.frames}

\item{config}{A named list. Needs to contain the component
'average_over_last_n_inj'.}
}
\value{
A named list of data.frames. The list elements are named like the
  input list "datasets". Each list element is a list with the collected
  output of \code{accumulateMeasurementsForSingleDataset()} and
  \code{getQualityControlInfo().}
}
\description{
Calculate the injection average of the samples and obtain the quality control
information based on the true standard values.
}
\details{
Uses the config parameter 'average_over_last_n_inj'. If it is
-1 or 'all', all injections are used to calculate the sample averages.
}
