% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIn.R
\name{ReadDataFiles}
\alias{ReadDataFiles}
\title{Read in Picarro data}
\usage{
ReadDataFiles(file, meta.info_folder)
}
\arguments{
\item{file}{character string with the path to the Picarro file which shall
be read.}

\item{meta.info_folder}{the output from \code{\link{get.input_folder}}.}
}
\value{
a list strucure with the file data and meta information.
}
\description{
Read in Picarro data from file.
}
\author{
Thomas Münch
}
