% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIn.R
\name{ReadStandardData}
\alias{ReadStandardData}
\title{Read isotope standard data}
\usage{
ReadStandardData(file_data, meta.info_std)
}
\arguments{
\item{file_data}{a data frame of Picarro file data.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}
}
\value{
a list structure with the standard isotope data.
}
\description{
Extract the isotope data of the standard samples from a data frame of read
Picarro file data.
}
\author{
Thomas Münch
}
