% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{ReturnInjNoData}
\alias{ReturnInjNoData}
\title{Data of a specific injection}
\usage{
ReturnInjNoData(x, sample_data)
}
\arguments{
\item{x}{integer; the injection number for which to extract the data.}

\item{sample_data}{data frame with the sample data from a certain measurement
file.}
}
\value{
a data frame with the extracted data.
}
\description{
Obtain all sample data for a given injection number.
}
\author{
Thomas Münch
}
