% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift-correction.R
\name{runDriftModel}
\alias{runDriftModel}
\title{Run simple drift estimation model}
\usage{
runDriftModel(trainingData, species = "d18O")
}
\arguments{
\item{trainingData}{a named list of input measurement datasets, where each
dataset are the measurement data of a specific standard which is to be
used for the drift estimation analysis. The names of the list should
correspond to the names of the used standards.}

\item{species}{character string with the name of the isotope species for
which the drift estimation model shall be calculated; valid names are
"d18O" for oxygen isotopes and "dD" for hydrogen isotopes.}
}
\value{
A tibble with \code{n + 1} rows, where \code{n} is the number of used
  drift monitoring standards and the additional row corresponds to the mean
  estimate across all drift standards, and with six variables:
\describe{
\item{\code{species}:}{character; the name of the isotope \code{species}
  used;}
\item{\code{sample}:}{the name of the used standard, or "mean" for the
  average value across all analysed standards;}
\item{\code{slope}:}{numeric; the estimated drift correction slope in permil
  per second;}
\item{\code{pValueSlope}:}{numeric; the p-value of the drift correction
  slope (\code{NA} for the mean estimate);}
\item{\code{residualRMSD}:}{numeric; the root mean square deviation of the
  drift regression residuals (\code{NA} for the mean estimate);}
\item{\code{rSquared}:}{numeric; the r-squared value of the drift
  regression (\code{NA} for the mean estimate).}
}
}
\description{
Run the simple drift estimation model of linearly regressing measured
standard values against the elapsed measurement time.
}
