% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{SampleCalibration}
\alias{SampleCalibration}
\title{Calibration of sample data}
\usage{
SampleCalibration(sample_data, calibration.results, flag.Calibration)
}
\arguments{
\item{sample_data}{a data frame with only the data of the samples for a
specific measurement file.}

\item{calibration.results}{the output from
\code{\link{wrapper.StandardCalibration}}.}

\item{flag.Calibration}{integer; flag to signal which calibration scheme
shall be applied (1 = linear drift correction + single three-point
calibration at the beginning of the measurement sequence; 2 = double
three-point calibration at the beginning and at the end of the measurement
sequence).}
}
\value{
as \code{sample_data} but with the calibrated data.
}
\description{
Function to perform a linear calibration of the measured sample isotope
data including a drift correction.
}
\author{
Thomas Münch
}
