% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{selectGroupsForTwoPointCalib}
\alias{selectGroupsForTwoPointCalib}
\title{Select lowest and highest standard}
\usage{
selectGroupsForTwoPointCalib(groups)
}
\arguments{
\item{groups}{a list of vectors from splitting a data frame block of
standards by their \code{Identifier 1} values through calling
\code{\link{split}}.}
}
\value{
A data frame with all injections from the two selected standards.
}
\description{
From a group of standards, select the two standards that exhibit the lowest
and highest isotope values.
}
