% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibFunctions.R
\name{StackCalibData}
\alias{StackCalibData}
\title{Output file data}
\usage{
StackCalibData(std_data, sample_data)
}
\arguments{
\item{std_data}{a list structure with the corrected and calibrated
standard data.}

\item{sample_data}{a data frame with the corrected and calibrated
sample data.}
}
\value{
a data frame with both data types merged together in the same format
as the input data files.
}
\description{
Convert corrected and calibrated standard and sample data back to a data
frame for output.
}
\author{
Thomas Münch
}
