% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{StandardCalibration}
\alias{StandardCalibration}
\title{Calibration of standard data}
\usage{
StandardCalibration(std_data, meta.info_std, flag.iso_type,
  flag.Calibration)
}
\arguments{
\item{std_data}{a data frame with only the data of the last three injections
of the standard samples for a specific measurement file.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}

\item{flag.iso_type}{character string of the isotopic species for which the
calibration is to be applied; either "delta.O18" or "delta.H2".}

\item{flag.Calibration}{integer; flag to signal which calibration scheme
shall be applied (1 = linear drift correction + single three-point
calibration at the beginning of the measurement sequence; 2 = double
three-point calibration at the beginning and at the end of the measurement
sequence).}
}
\value{
a list structure with the calibrated standard data along with
statistical information on the quality of the calibration approach.
}
\description{
Function to perform a linear calibration of the measured standard isotope
data for a specific isotope species given literature delta values for each
standard.
}
\author{
Thomas Münch
}
