% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataHandling.R
\name{TakeMean_CalcPooledSD.sample_data}
\alias{TakeMean_CalcPooledSD.sample_data}
\title{Sample means and pooled standard deviations}
\usage{
TakeMean_CalcPooledSD.sample_data(sample_data)
}
\arguments{
\item{sample_data}{a data frame with the sample data for a specific
measurement file.}
}
\value{
a two element list: (1) a data frame similar to \code{sample_data}
with the mean sample data of the last three injections; (2) an array with
the pooled standard deviations.
}
\description{
Function to calculate the mean and the pooled standard deviation over the
last three injections of the sample data.
}
\author{
Thomas Münch
}
