% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MemoryCorrection.R
\name{wrapper.MemCorrStdData}
\alias{wrapper.MemCorrStdData}
\title{Apply memory correction}
\usage{
wrapper.MemCorrStdData(x, std_data, file_data, mem.coeff_O18, mem.coeff_H2,
  meta.info_std)
}
\arguments{
\item{x}{the number of the standard to be corrected.}

\item{std_data}{subset of \code{file_data} with only the data of the
standard samples.}

\item{file_data}{data frame of raw isotope data of a specific measurement
file.}

\item{mem.coeff_O18}{numeric vector of memory coefficients for oxygen
isotope data (d18O).}

\item{mem.coeff_H2}{numeric vector of memory coefficients for deuterium
isotope data (d2H).}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}
}
\value{
the memory-corrected subset of the standard data of standard
no. \code{x}.
}
\description{
Apply the memory correction scheme of van Geldern and Barth (2012) on
standard data.
}
\author{
Thomas Münch
}
