% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIn.R
\name{wrapper.ReadDataFiles}
\alias{wrapper.ReadDataFiles}
\title{Read in Picarro data}
\usage{
wrapper.ReadDataFiles(meta.info_folder)
}
\arguments{
\item{meta.info_folder}{the output from \code{\link{get.input_folder}}.}
}
\value{
a list structure with the Picarro data from all files.
}
\description{
Wrapper function to read in Picarro data from all files in a given folder.
}
\author{
Thomas Münch
}
