% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{wrapper.StandardCalibration}
\alias{wrapper.StandardCalibration}
\title{Prepare standard calibration}
\usage{
wrapper.StandardCalibration(std_data, meta.info_std, flag.Calibration)
}
\arguments{
\item{std_data}{a data frame with only the data of the standard samples for
a specific measurement file.}

\item{meta.info_std}{the respective sublist of the output from
\code{\link{get.input_folder}}.}

\item{flag.Calibration}{integer; flag to signal which calibration scheme
shall be applied (1 = linear drift correction + single three-point
calibration at the beginning of the measurement sequence; 2 = double
three-point calibration at the beginning and at the end of the measurement
sequence).}
}
\value{
a list structure with the calibrated standard data along with
statistical information on the quality of the calibration and drift
correction approach.
}
\description{
Wrapper function for the drift-correction and the linear calibration of the
standard isotope data.
}
\author{
Thomas Münch
}
