% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{WriteOutput}
\alias{WriteOutput}
\title{Write "piccr" output}
\usage{
WriteOutput(calibration.results, meta.info_folder)
}
\arguments{
\item{calibration.results}{the output from
\code{\link{CalibratePicarroData}}.}

\item{meta.info_folder}{the output from \code{\link{get.input_folder}}.}
}
\description{
Write the corrected and calibrated isotope data from all read files to disk,
alongside with statistical information on the quality of the
correction/calibration approach.
}
\author{
Thomas Münch
}
