% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{writeSingleDataset}
\alias{writeSingleDataset}
\title{Write a data set to disc}
\usage{
writeSingleDataset(dataset, folder, config)
}
\arguments{
\item{dataset}{a data frame of processed measurement data.}

\item{folder}{character string with the directory path in which to save the
data.}

\item{config}{a named list containing the component
\code{include_standards_in_output}: a logical to signal whether to save
the processed data including the measured standards or not.}
}
\description{
Write the given processed data set to disc as a csv file in the specified
directory.
}
