% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_processDataForOutput.R
\name{accumulateMeasurements}
\alias{accumulateMeasurements}
\title{accumulate measurements}
\usage{
accumulateMeasurements(dataset, config)
}
\arguments{
\item{config}{A named list. Needs to contain the component
'average_over_last_n_inj'.}

\item{datasets}{A data frame with the data set.}
}
\value{
A data frame.
}
\description{
Average the delta.O18, delta.H2 and d.Excess values for each 
sample from a dataset and calculate their standard deviation.
}
\details{
Uses the config parameter 'average_over_last_n_inj'. If it is
-1 or 'all', all injections are used to calculate the averages
and standard deviations.
}
