% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_addColumnDExcess.R
\name{addColumnDExcess}
\alias{addColumnDExcess}
\title{addColumnDExcess}
\usage{
addColumnDExcess(datasets)
}
\arguments{
\item{datasets}{A named list of dataframes.}
}
\value{
A list. The list elements are named like the input list "datasets". 
        Each element of the list is a dataframe that includes the column
        "dExcess".
}
\description{
Add the column "dExcess" to every dataframe in datasets. dExcess is 
calculated as dExcess = dD - 8 * d18O.
}
