% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_addColumnDExcess.R
\name{addColumnDExcess}
\alias{addColumnDExcess}
\title{Add column dExcess}
\usage{
addColumnDExcess(dataset)
}
\arguments{
\item{dataset}{A data.frame.}
}
\value{
A data.frame.
}
\description{
Add the column "dExcess" to the dataset. dExcess is 
calculated as dExcess = dD - 8 * d18O.
}
