% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{addColumnSecondsSinceStart}
\alias{addColumnSecondsSinceStart}
\title{Add the column 'SecondsSinceStart' to the given dataset.}
\usage{
addColumnSecondsSinceStart(dataset)
}
\arguments{
\item{dataset}{A data.frame. Needs to contain the column
'Time Code'. The elements of 'Time Code' should
be character vectors of the format 'yyyy/mm/ddhh:mm:ss'
(e.g. '2019/11/2510:00:00').}
}
\value{
A data.frame that includes the column 'SecondsSinceStart'.
}
\description{
Add the column 'SecondsSinceStart' to the given dataset.
}
