% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{applyCalibration}
\alias{applyCalibration}
\title{Apply linear calibration}
\usage{
applyCalibration(dataset, calibrationParams)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{calibrationParams}{a named list with elements \code{d18O} and \code{dD}
  where each element is again a list with two elements:
  \describe{
  \item{\code{intercept}:}{numeric; calibration intercept.}
  \item{\code{slope}:}{numeric; calibration slope.}
}}
}
\value{
The input \code{dataset} with the d18O and dD values calibrated
  according to the given calibration parameters.
}
\description{
Apply a linear calibration to a specific data set given calibration slope and
intercept values.
}
