% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{applyCalibration}
\alias{applyCalibration}
\title{Apply linear calibration}
\usage{
applyCalibration(dataset, calibrationParams)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{calibrationParams}{the calibration parameters in a tibble with two rows
  and three mandatory variables:
  \describe{
  \item{\code{species}:}{character; must be \code{d18O} for one row and
    \code{dD} for the other.}
  \item{\code{intercept}:}{numeric; the respective calibration intercept.}
  \item{\code{slope}:}{numeric; the respective calibration slope.}
}}
}
\value{
The input \code{dataset} with the d18O and dD values calibrated
  according to the given calibration parameters.
}
\description{
Apply a linear calibration to a specific data set given calibration slope and
intercept values.
}
