% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingDoubleCalibration.R
\name{applyDoubleCalibration}
\alias{applyDoubleCalibration}
\title{Apply double-block calibration}
\usage{
applyDoubleCalibration(dataset, calibParams)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{calibParams}{a tibble with the calibration parameters (slope and
intercept) for d18O and dD (see \code{\link{getCalibration}} for details on
the tibble structure) as estimated from the first standard block in the
measurement sequence as well as from the last standard block.}
}
\value{
The input \code{dataset} with the d18O and dD values calibrated
  according to the time-varying calibration slope and intercept.
}
\description{
Apply a double-block calibration to a specific data set using provided
calibration parameters from the first and last standard block in the
measurement sequence, which are linearly interpolated in between.
}
