% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingDoubleCalibration.R
\name{applyDoubleCalibration}
\alias{applyDoubleCalibration}
\title{Apply double-block calibration}
\usage{
applyDoubleCalibration(dataset, calibParamsBlock1, calibSlopes)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{calibParamsBlock1}{a set of calibration parameters (slope and
intercept) for d18O and dD (see \code{\link{getCalibInterceptAndSlope}} for
details) as estimated from the first standard block in the measurement
sequence.}

\item{calibSlopes}{slope estimates for d18O and dD for a linear change of the
calibration parameters in \code{calibParamsBlock1} across the mesasurement
sequence (see \code{\link{getCalibrationSlopes}} for details).}
}
\value{
The input \code{dataset} with the d18O and dD values calibrated
  according to a time-varying calibration slope and intercept.
}
\description{
Apply a double-block calibration to a specific data set using provided
calibration parameters.
}
