% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{applyDriftCorrection}
\alias{applyDriftCorrection}
\title{Apply drift correction}
\usage{
applyDriftCorrection(dataset, alpha)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{alpha}{numeric; a constant drift slope in permil per seconds.}
}
\value{
The input \code{dataset} with the d18O and dD values corrected
  against a linear drift with slope \code{alpha}.
}
\description{
Apply a linear drift correction to a specific data set given a constant
drift slope.
}
