% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_correctForMemoryEffect.R
\name{applyMemoryCorrection}
\alias{applyMemoryCorrection}
\title{Apply memory correction}
\usage{
applyMemoryCorrection(dataset, memoryCoefficients)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{memoryCoefficients}{a data frame of memory coefficients; must include
the columns \code{Inj Nr} (injection number), \code{memoryCoeffD18O}
(memory coefficients for d18O) and \code{memoryCoeffDD} (memory
coefficients for dD).}
}
\value{
The input \code{dataset} with the d18O and dD columns corrected for
  memory effect.
}
\description{
Apply a memory correction to a specific data set given a set of memory
coefficients for d18O and dD as a function of the injection number.
}
\details{
Note that if the number of supplied memory coefficients does not match the
maximum number of sample injections in the given data set, the memory
coefficients are padded with \code{1}'s to match the data. The memory
correction is applied according to the formula in
\code{\link{formulaCorrectMem}}.
}
\seealso{
\code{\link{formulaCorrectMem}}
}
