% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{associateStandardsWithConfigInfo}
\alias{associateStandardsWithConfigInfo}
\title{Associate standards with config info}
\usage{
associateStandardsWithConfigInfo(datasets, config)
}
\arguments{
\item{datasets}{A named list of dataframes.}

\item{config}{A named list. Needs to contain the component
'standards'; a named list with the components
'name', 'o18_True', 'H2_True', 'use_for_drift_correction',
'use_for_calibration', and 'use_as_control_standard'.}
}
\value{
A named list of dataframes. The names are equal to the names
        of the input 'datasets'.
}
\description{
Associate each standard in the given datasets with the 
config info for that standard (true values for 018 
and H2, use for drift correction, use for calibration).
}
\details{
The output datasets will contain the columns o18_True, H2_True,
useForDriftCorrection, and useForCalibration. The values will
be NA for probes.
}
