% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{buildOutputList}
\alias{buildOutputList}
\title{Build output list}
\usage{
buildOutputList(name, config, dataset, memoryCorrected, memoryCoefficients,
  calibrated, calibratedAndDriftCorrected, accumulated, qualityControlInfo)
}
\arguments{
\item{config}{A list. Needs to contain the component $use_memory_correction, a logical.}

\item{dataset}{A data.frame.}

\item{memoryCorrected}{A data.frame.}

\item{memoryCoefficients}{A data.frame.}

\item{calibrated}{A data.frame.}

\item{calibratedAndDriftCorrected}{A data.frame.}

\item{processedData}{A named list. Has the components $accumulatedData (a data.frame), 
$deviationsFromTrue (a data.frame), $rmsdDeviationsFromTrue
(a list), $pooledSD (a list), and $deviationOfControlStandard (a list).}
}
\value{
A nested list structure
}
\description{
This is the final processing step for each dataset. After all the processing is done, input
all relevant information and construct the output for a single dataset for the function
processData(..).
}
\details{
Note: If you change the output of this function, don't forget to update
the roxygen docstring (section return value) of the function processData(..).
}
