% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{buildSecondSection}
\alias{buildSecondSection}
\title{Build section "VALUES FOR EACH FILE"}
\usage{
buildSecondSection(processedData)
}
\description{
Build section "VALUES FOR EACH FILE"
}
