% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{calculateDriftSlope}
\alias{calculateDriftSlope}
\title{Calculate linear drift slope}
\usage{
calculateDriftSlope(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{useForDriftCorr} and
\code{vial_group}, which are not included in the raw Picarro output.}

\item{config}{a named list with the logical component
  \code{use_memory_correction} to signal whether a memory correction has been
applied to the input data.}
}
\value{
A named list with elements \code{d18O} and \code{dD} which contain
  the average drift slope in permil per seconds for d18O and dD.
}
\description{
Calculate a linear drift slope for each isotope species from the average
across individual estimated drift slopes from the drift monitoring standards
specified in the configuration. If memory correction was applied to the data,
all injections are used from the drift monitoring standards, else only the
last three injections.
}
\seealso{
\code{\link{associateStandardsWithConfigInfo}},
  \code{\link{assignVialsToGroups}}
}
