% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calculatePoooledStdDev.R
\name{calculatePoooledSD}
\alias{calculatePoooledSD}
\title{calculate poooled standard deviation}
\usage{
calculatePoooledSD(dataset)
}
\arguments{
\item{datasets}{A data.frame with isotope measurement data.}
}
\value{
A list with the two named elements
        "d18O" and "dD" (the pooled standard deviation for d18O and dD).
}
\description{
Calculate the pooled standard deviation for d18O and dD. The pooled standard
deviation is calculated using the following formulas:
}
\details{
x = (n_1 - 1) \* sd_1^2 + ... + (n_k - 1) \* sd_k^2
    y = n_1 + ... + n_k - k
    sdPooled = sqrt( x / y )
}
