% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calculateRMSD.R
\name{calculateRMSD}
\alias{calculateRMSD}
\title{Calculate root-mean-square deviation}
\usage{
calculateRMSD(v1, v2, na.rm = FALSE)
}
\arguments{
\item{v1}{numeric vector for which to compute the rmsd with \code{v2}.}

\item{v2}{numeric vector for which to compute the rmsd with \code{v1}; must
be of the same length as \code{v1}.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
The root-mean-square deviation of \code{v1} and \code{v2}, or
\code{NA} (for \code{na.rm = FALSE}) if any of their elements is \code{NA}.
}
\description{
Calculate the root-mean-square deviation (rmsd) of two numeric vectors.
}
