% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingDoubleCalibration.R
\name{calibrateUsingDoubleCalibration}
\alias{calibrateUsingDoubleCalibration}
\title{Calibrate the given dataset using double calibration}
\usage{
calibrateUsingDoubleCalibration(dataset, config)
}
\arguments{
\item{dataset}{A data.frame with isotope measurement data.
Should contain the additional columns "block" and 
"useForMemCorr" (not included in the raw Picarro output).}

\item{config}{A named list. Need to contain the boolean elements "use_memory_correction"
and "use_three_point_calibration".}
}
\value{
A data.frame.
}
\description{
Calibrate the given dataset using double calibration
}
