% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingDoubleCalibration.R
\name{calibrateUsingDoubleCalibration}
\alias{calibrateUsingDoubleCalibration}
\title{Calibrate data using double-block calibration}
\usage{
calibrateUsingDoubleCalibration(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{o18_True} and \code{H2_True} which are not
included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}
}
\value{
A named list with two elements:
\describe{
\item{dataset:}{the input \code{dataset} with the d18O and dD values
  calibrated according to the double-block calibration.}
\item{parameter:}{a tibble with the applied calibration parameters, and their
  associated quality control information; i.e. the output from
  \code{\link{getCalibration}} for the first and the final standard block of
  the measurement sequence bound together in a single tibble.}
}
}
\description{
Calibrate a given data set using calibration slope and intercept values which
are linearly interpolated between the first and the final block of standard
measurements within the measurement sequence.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}}
}
