% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{calibrateUsingSimpleDriftCorrection}
\alias{calibrateUsingSimpleDriftCorrection}
\title{Calibrate data using single-block calibration and linear drift correction}
\usage{
calibrateUsingSimpleDriftCorrection(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{useForDriftCorr}, \code{o18_True} and
\code{H2_True} which are not included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}
}
\value{
A named list with three elements:
\describe{
\item{dataset:}{the input \code{dataset} with the d18O and dD values
  calibrated according to linear drift correction and single calibration.}
\item{calibrationParameter:}{a tibble with the applied calibration
  parameters, and their associated quality control information, output from
  \code{\link{getCalibration}}.}
\item{driftParameter:}{a tibble with the applied drift correction parameters,
  and their associated quality control information, output from
  \code{\link{calculateDriftSlope}}.}
}
}
\description{
Calibrate a given data set using calibration slope and intercept values
estimated with standards from the beginning of the measurement sequence
following a linear drift correction against measurement time.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}},
  \code{\link{linearCalibration}},
  \code{\link{linearDriftCorrection}},
  \code{\link{getCalibration}},
  \code{\link{calculateDriftSlope}}
}
