% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{calibrateUsingSimpleDriftCorrection}
\alias{calibrateUsingSimpleDriftCorrection}
\title{Calibrate using simple drift correction}
\usage{
calibrateUsingSimpleDriftCorrection(dataset, config)
}
\arguments{
\item{config}{A named list of config arguments. The required config arguments 
are the same as for the function "linearDriftCorrection".}

\item{datasets}{A data.frame with isotope measurement data.
Should contain the additional columns "block", 
"useForDriftCorr", and "useForMemCorr" (not included in the 
raw Picarro output).}
}
\value{
A data.frame.
}
\description{
Apply calibration and drift correction to the input dataset. Uses
linear drift correction and linear calibration. The calibration uses
only the standards in block 1.
}
