% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{calibrateUsingSimpleDriftCorrection}
\alias{calibrateUsingSimpleDriftCorrection}
\title{calibrateUsingSimpleDriftCorrection}
\usage{
calibrateUsingSimpleDriftCorrection(datasets, config)
}
\arguments{
\item{datasets}{A named list of dataframes with isotope measurement data.
Each dataframe should contain the additional columns "block", 
"useForDriftCorr", and "useForMemCorr" (not included in the 
raw Picarro output).}

\item{config}{A named list of config arguments. The required config arguments 
are the same as for the function "linearDriftCorrection".}
}
\value{
A list. The list elements are named like the input list "datasets". 
        Each element of the list is a dataframe with data that
        has been calibrated and drift corrected.
}
\description{
Apply calibration and drift correction to the input datasets. Uses
linear drift correction and linear calibration. The calibration uses
only the standards in block 1.
}
