% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_correctForMemoryEffect.R
\name{correctForMemoryEffect}
\alias{correctForMemoryEffect}
\title{Correct for memory effect}
\usage{
correctForMemoryEffect(dataset)
}
\arguments{
\item{dataset}{A data.frame.}
}
\value{
A list. It contains the two named elements
        "datasetMemoryCorrected" and "memoryCoefficients".
}
\description{
Take a data.frame with isotope measurement data and apply memory correction to it.
}
