% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_processDataForOutput.R
\name{filterInjections}
\alias{filterInjections}
\title{Filter injection range}
\usage{
filterInjections(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{config}{A named list of configuration parameters (e.g. as read from the
\code{config.yaml} file) containing at least the component
\code{average_over_inj}.}
}
\value{
The input data frame \code{dataset} with only the specified
injections remaining.
}
\description{
Filter out a specified injection range from a measurement data frame.
}
\details{
Filtering is done according to the configuration parameter
\code{average_over_inj} (e.g. see the \code{config.yaml} file). If the
parameter is set to \code{all} or -1, all injections are kept, if it is set
to a single integer `n`, the last `n` injections are kept, else if it is set
to a range `n1:n2`, injections `n1:n2` are kept.
}
