% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_correctForMemoryEffect.R
\name{formulaCorrectMem}
\alias{formulaCorrectMem}
\title{Memory correction}
\usage{
formulaCorrectMem(data, memCoeff, deltaTruePrev)
}
\arguments{
\item{data}{numeric vector with measured sample values.}

\item{memCoeff}{numeric vector with memory coefficients matching the
injection number of the values in \code{data}.}

\item{deltaTruePrev}{numeric vector of (estimated) true values of the
respective previous sample of the samples in \code{data}.}
}
\value{
the input vector \code{data} corrected for memory.
}
\description{
Formula for the memory correction; see the package vignette for
details.
}
