% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_correctForMemoryEffect.R
\name{formulaMemCoeff}
\alias{formulaMemCoeff}
\title{Memory coefficients}
\usage{
formulaMemCoeff(data, deltaTrue, deltaTruePrev)
}
\arguments{
\item{data}{numeric vector with measured values of a certain sample for
consecutive injections.}

\item{deltaTrue}{numeric vector with the (estimated) true value of that
sample.}

\item{deltaTruePrev}{numeric vector with the (estimated) true value of the
previous sample.}
}
\value{
numeric vector of the same length as \code{data} with the memory
  coefficients.
}
\description{
Formula to calculate memory coefficients; see the package vignette for
details.
}
