% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{getCalibration}
\alias{getCalibration}
\title{Get calibration parameters}
\usage{
getCalibration(dataset, config, useBlock)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{o18_True} and \code{H2_True} which are not
included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}

\item{useBlock}{a single integer giving the number of the standard block which
is to be used for estimating the calibration parameters.}
}
\value{
A named list with elements \code{d18O} and \code{dD} where each
element is again a list containing the output of
  \code{\link{runCalibrationModel}}.
}
\description{
Obtain the calibration parameters as estimated from standards in a particular
standard block of the measurement sequence from regressing the expected
standard values against their measured values; for this, a simple linear
regression is currently implemented.
}
\details{
If memory correction was applied to the data, all injections are used from
the standard data, else only the last three injections. If a two-point
calibration is switched on, only the two standards with the lowest and
highest isotopic values are used, regardless of how many calibration
standards are actually set in the dataset.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}},
  \code{\link{runCalibrationModel}}
}
