% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingDoubleCalibration.R
\name{getCalibrationSlopes}
\alias{getCalibrationSlopes}
\title{Temporal change of calibration parameters}
\usage{
getCalibrationSlopes(paramsBlock1, paramsBlockN)
}
\arguments{
\item{paramsBlock1}{the calibration parameters estimated from the first
standard block.}

\item{paramsBlockN}{the calibration parameters estimated from the final
standard block.}
}
\value{
A named list with elements \code{d18O} and \code{dD} where each
element is again a list with two elements:
\describe{
\item{\code{alpha}:}{numeric; the slope of the linear change in the
  calibration intercept across the measurement sequence.}
\item{\code{beta}:}{numeric; the slope of the linear change in the
  calibration slope across the measurement sequence.}
}
}
\description{
Estimate the change in calibration parameters between the beginning and the
end of the measurement sequence based on a simple linear regression of the
calibration parameters against the elapsed measurement time.
}
