% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_correctForMemoryEffect.R
\name{getDeltaTrueAndDeltaTruePrevForEachSample}
\alias{getDeltaTrueAndDeltaTruePrevForEachSample}
\title{Estimate true sample values}
\usage{
getDeltaTrueAndDeltaTruePrevForEachSample(dataset, ...)
}
\arguments{
\item{dataset}{a data frame with a block of standard vial measurements;
usually the first standard block at the beginning of the measurement
sequence.}

\item{...}{names of column variables which are used to group the data in
\code{block} into unique sets of standard vials; in the default programme
flow of \code{piccr}, \code{Identifier 1} and \code{vial_group} are used.}
}
\value{
The input data set supplemented by four new columns:
  \describe{
  \item{\code{deltaTrueD18O}:}{the estimated true d18O value of the current
    vial.}
  \item{\code{deltaTrueDD}:}{the estimated true dD value of the current
    vial.}
  \item{\code{deltaTruePrevD18O}:}{the estimated true d18O value of the
    previous vial.}
  \item{\code{deltaTruePrevDD}:}{the estimated true dD value of the previous
    vial.}
}
}
\description{
To calculate memory coefficients, an estimate of the true values of the
standard vials is needed. This function estimates these from the mean of the
last three injections of the standard vials, thereby assuming a sufficiently
large number of injections to provide good, i.e. memory-free, estimates.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{assignVialsToGroups}}
}
