% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_accumulateMeasurementsForEachSample.R
\name{getQualityControlInfo}
\alias{getQualityControlInfo}
\title{Obtain quality control information}
\usage{
getQualityControlInfo(dataset, accumulatedDataset)
}
\arguments{
\item{dataset}{a data frame with corrected and calibrated measurement data
of a specific data set.}

\item{accumulatedDataset}{a data frame with the accumulated
(i.e. injection-averaged) measurement data for this data set.}
}
\value{
A list with three elements:
  $deviationsFromTrue (data frame with the deviations from the true value for
   each measured standard.)
  $rmsdDeviationsFromTrue (list with elements \code{d18O} and \code{dD} with
   the rmsd across \code{deviationsFromTrue} (very first standard excluded).)
  $deviationOfControlStandard (list with elements \code{Identifier 1},
  \code{d18O} and \code{dD} with the deviations from the true value for the
  quality control standard(s) in \code{Identifier 1}.)
}
\description{
Obtain the quality control information for a data set based on the deviations
of the measured standards from their true values.
}
