% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_processDataForOutput.R
\name{getQualityControlInfo}
\alias{getQualityControlInfo}
\title{Obtain quality control information}
\usage{
getQualityControlInfo(dataset, accumulatedDataset)
}
\arguments{
\item{dataset}{a data frame with corrected and calibrated measurement data of
a specific data set.}

\item{accumulatedDataset}{a data frame with the accumulated
(i.e. injection-averaged) measurement data for this data set.}
}
\value{
A list with three elements:
  \describe{
  \item{\code{deviationsFromTrue}:}{data frame with the measured (and
    processsed) value (d18O and dH), the expected value, and the deviation of
    the measured from the expected value for each analysed standard.}
  \item{\code{rmsdDeviationsFromTrue}:}{a list with the two elements
    \code{d18O} and \code{dD} with the respective root mean square deviation
    across all deviations in \code{deviationsFromTrue} (with the very first
    standard of the measurement sequence excluded).}
  \item{\code{deviationOfControlStandard}:}{a list with the two elements
    \code{d18O} and \code{dD} of the deviations from the expected value for
    the quality control standard(s) specified in the config file.}
  \item{\code{pooledSD}:}{a list with the two elements \code{d18O}
    and \code{dD} of the pooled standard deviations across all measured
    vials.}
}
}
\description{
Obtain the quality control information for a processed data set based on the
deviations of the measured standards from their expected values.
}
\seealso{
\code{\link{calculatePoooledSD}}
}
