% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_processDataForOutput.R
\name{getVialCountOfFirstStd}
\alias{getVialCountOfFirstStd}
\title{Number of warm-up standard vials}
\usage{
getVialCountOfFirstStd(dataset)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}
}
\value{
A single integer with the number of the last vial of the very
first standard in the input \code{dataset}.
}
\description{
Get the number of vials the very first standard in the measurement sequence
(the "warm-up standard") is injected from, i.e. "vial grouping" is accounted
for, so if the first standard is injected from several vials in a row, the
number of the last of these vials is returned.
}
