% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{groupStandardsInBlocks}
\alias{groupStandardsInBlocks}
\title{Group standards in blocks}
\usage{
groupStandardsInBlocks(dataset, config)
}
\arguments{
\item{dataset}{A data.frame.}

\item{config}{A named list. Needs to contain the component
$standards; a list of lists where each innermost
list needs to contain the component $name.}
}
\value{
A data.frame.
}
\description{
For each standard injection, determine which standard block
it belongs to. The results are stored in the column 'block'.
For probes the value is NA.
}
