% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{groupStandardsInBlocks}
\alias{groupStandardsInBlocks}
\title{Group standards in blocks}
\usage{
groupStandardsInBlocks(datasets, config)
}
\arguments{
\item{datasets}{A named list of dataframes.}

\item{config}{A named list. Needs to contain the component
$standards; a list of lists where each innermost
list needs to contain the component $name.}
}
\value{
A named list of dataframes. The names are the same as
        the names of the input 'datasets'.
}
\description{
For each standard injection, determine which standard block
it belongs to. The results are stored in the column 'block'.
For probes the value is NA.
}
