% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{isStandard}
\alias{isStandard}
\title{Determine if a given ID belongs to a standard.}
\usage{
isStandard(id1, config)
}
\arguments{
\item{id1}{A character vector. The id to test for.}

\item{config}{A named list. Needs to contain the component 'standards'.
'standards' is a list, each element needs to contain the
component 'name'.}
}
\value{
A logical.
}
\description{
Determine if a given ID belongs to a standard.
}
