% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{linearCalibration}
\alias{linearCalibration}
\title{Calibrate data using single block}
\usage{
linearCalibration(dataset, config, block = 1)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{block},
\code{useForCalibration}, \code{o18_True} and \code{H2_True} which are not
included in the raw Picarro output.}

\item{config}{a named list which needs to contain the following elements:
\describe{
  \item{\code{use_memory_correction}:}{logical; has a memory correction been
  applied to the input data?}
  \item{\code{use_three_point_calibration}:}{logical; shall three or more
    standards as specified by the data set column \code{useForCalibration} be
    used as calibration standards (\code{TRUE}) or only two (\code{FALSE})?}
}}

\item{block}{a single integer giving the number of the standard block which
is to be used for estimating the calibration parameters; defaults to
\code{1} (first block in the measurement sequence).}
}
\value{
A named list with two elements:
\describe{
\item{dataset:}{the input \code{dataset} with the d18O and dD values
  calibrated according to the single-block calibration.}
\item{parameter:}{a tibble with the applied calibration parameters, and their
  associated quality control information, output from
  \code{\link{getCalibration}}.}
}
}
\description{
Calibrate a given data set using calibration slope and intercept values
estimated with standards from a particular standard block of the measurement
sequence. The calibration parameters are obtained from regressing the
expected standard values against their measured values.
}
\details{
If memory correction was applied to the data, all injections are used from
the standard data, else only the last three injections. If a two-point
calibration is switched on, only the two standards with the lowest and
highest isotopic values are used, regardless of how many calibration
standards are actually set in the dataset.
}
\seealso{
\code{\link{groupStandardsInBlocks}},
  \code{\link{associateStandardsWithConfigInfo}}
}
