% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{linearCalibration}
\alias{linearCalibration}
\title{linearCalibration}
\usage{
linearCalibration(dataset, config, block = 1)
}
\arguments{
\item{dataset}{A data.frame with isotope measurement data. It dataframe should 
contain the additional columns "block" and 
"useForMemCorr" (not included in the raw Picarro output).}

\item{config}{A named list. Need to contain the boolean elements "use_memory_correction"
and "use_three_point_calibration".}

\item{block}{A number. Use the standards in this block for calibration. Default: 1.}
}
\value{
A data.frame.
}
\description{
Take a data.frame with isotope measurement data and 
calibrate it. Does not include drift correction.
Note: 
 - If config$use_memory_correction is TRUE all injections are used for 
   calibration, else only the last three injections are used.
 - If config$use_triple_calibration is TRUE all standards in the specified
   block are used for calibration. If it is FALSE, only the first and last
   standards are used.
}
