% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_calibrateUsingSimpleDriftCorrection.R
\name{linearDriftCorrection}
\alias{linearDriftCorrection}
\title{Drift-correct data}
\usage{
linearDriftCorrection(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set. It
needs to contain the additional columns \code{useForDriftCorr} and
\code{vial_group}, which are not included in the raw Picarro output.}

\item{config}{a named list with the logical component
  \code{use_memory_correction} to signal whether a memory correction has been
applied to the input data.}
}
\value{
A named list with two elements:
\describe{
\item{dataset:}{the input \code{dataset} with the d18O and dD values
  corrected for a constant linear drift.}
\item{parameter:}{a tibble with the applied drift correction parameters, and
  their associated quality control information, output from
  \code{\link{calculateDriftSlope}}.}
}
}
\description{
Correct a given data set for an assumed constant linear drift of the isotope
values with measurement time.
}
\seealso{
\code{\link{associateStandardsWithConfigInfo}},
  \code{\link{assignVialsToGroups}},
  \code{\link{calculateDriftSlope}}
}
