% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{normalizeInjectionNumbers}
\alias{normalizeInjectionNumbers}
\title{Re-calculate injection numbers}
\usage{
normalizeInjectionNumbers(datasets)
}
\arguments{
\item{datasets}{A named list of data frames; each data frame is a read-in
data set.}
}
\value{
The same named list of data frames where for each data set the
        column of injection numbers has been re-calculated accounting for
        possible consecutive vials of the same probe.
}
\description{
Re-calculate injection numbers to account for probes being measured
from two or more consectuive vials.
}
