% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{normalizeInjectionNumbers}
\alias{normalizeInjectionNumbers}
\title{Re-calculate injection numbers}
\usage{
normalizeInjectionNumbers(dataset)
}
\arguments{
\item{datasets}{A data.frame.}
}
\value{
A data.frame. The column of injection numbers has been 
        re-calculated accounting for possible consecutive vials 
        of the same probe.
}
\description{
Re-calculate injection numbers to account for probes being measured
from two or more consectuive vials.
}
