% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{outputSummaryFile}
\alias{outputSummaryFile}
\title{Output quality control summary file}
\usage{
outputSummaryFile(processedData, config, outputFile = NULL)
}
\arguments{
\item{processedData}{the processed measurement data as output by
\code{\link{processData}}.}

\item{config}{a named list with the contents read from the YAML configuration
file, which was used for the processing; at least, it needs to contain the
element \code{output_directory} to set the output directory for the summary
file, if no explicit directory is given in the argument \code{outputFile},
and the element \code{config_file_name} with the file path of the used YAML
configuration file.}

\item{outputFile}{an optional character vector with an alternative file path
(including file name) for the summary file. If \code{NULL} (the default),
the file \code{run.info} in the directory \code{config$output_directory}
is created.}
}
\description{
Write an output file to disc summarising the quality control information from
the processing steps for each processed file as well as for the entire
processing run. Note that if the data is processed again, the summary file is
silently overwritten.
}
\details{
The summary file consists of up to six sections:
\itemize{
  \item the first section gives the overall values for the entire processing
        run of the root-mean-square deviation (rmsd) of the quality control
        standards, the rmsd of all used standards, and the mean value of the
        pooled standard deviation per sample;
  \item the second section contains tables with the above values for each
        processed measurement file;
  \item the third section contains the estimated memory coefficients and
        their standard deviation, as a function of the injection number, for
        each processed measurement file as well as for the average across all
        processed files. This is only printed if a memory correction was
        actually applied to the measurement data.
  \item the fourth section contains information on the applied calibration
        regressions for each processed measurement file.
  \item the fifth section contains information on an applied linear drift
        correction for each measurement file. This is only printed if such a
        drift correction was actually applied to the measurement data.
  \item the sixth section contains a table which lists for each processed
        measurement file the individual deviations of each used standard from
        its true value.
}
}
\seealso{
\code{\link{processFiles}}
}
