% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{outputSummaryFile}
\alias{outputSummaryFile}
\title{outputSummaryFile}
\usage{
outputSummaryFile(processedData, config, outputFile = NULL)
}
\arguments{
\item{processedData}{A list as output by processData(..). 
It needs to contain the components
$pooledStdDev (a named list where each
element is of the form 'list(d18O = .., dD = ..)')
and $processed (a named list of dataframes).}

\item{config}{A named list. Needs to contain the component 
$standards (A list of lists. Each innermost list 
contains the components $name, $o18_True, and  $H2_True).
Should contain $output_directory if no explicit
value is given for the argument outputFile.}

\item{outputFile}{(Optional) A character vector. Path to the output file.
If it is not given, the file 'run.info' in the
directory config$output_directory is used.}
}
\value{
No relevant return value
}
\description{
Write a summary file to disc. The summary file contains
quality indicators and consists of three sections: 
  - AVERAGE OVER ALL FILES: This section contains the
    pooled standard deviation (O18 and H2) averaged 
    over all files.
    (More parameters will be added in the future.)
  - VALUES FOR EACH FILE: This section contains the 
    pooled standard deviation (O18 and H2) for each file.
  - INTER STANDARD BIAS TO LITERATURE VALUES FOR EACH FILE:
    This section displays the deviation of the measured
    delta.O18 and delta.H2 values to the true values for
    each file.
}
