% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{parseConfig}
\alias{parseConfig}
\title{Read the YAML configuration file}
\usage{
parseConfig(configFile)
}
\arguments{
\item{configFile}{a character string with the file path of the configuration
file.}
}
\value{
A list of the parameters read from the configuration file.
}
\description{
Read in the specified YAML configuration file for the \code{piccr}
processing.
}
