% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{parseConfig}
\alias{parseConfig}
\title{Read the specified YAML configuration file for piccr processing.}
\usage{
parseConfig(configFile)
}
\arguments{
\item{configFile}{A character string naming the config file.}
}
\value{
A list.
}
\description{
Read the specified YAML configuration file for piccr processing.
}
