% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_outputSummaryFile.R
\name{printQualityControl}
\alias{printQualityControl}
\title{Print quality control}
\usage{
printQualityControl(datasets, printDeviations = FALSE, n = 3)
}
\arguments{
\item{datasets}{dummy}

\item{printDeviations}{dummy}

\item{n}{dummy}
}
\value{
dummy
}
\description{
dummy text
}
