% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_processData.R
\name{processData}
\alias{processData}
\title{Process Picarro CRDS stable isotope data}
\usage{
processData(datasets, config)
}
\arguments{
\item{datasets}{a named list of data frames with the raw isotope measurement
data as output by the Picarro device.}

\item{config}{a named list of the following required components:
\describe{
\item{\code{use_memory_correction}.}{logical; shall a memory correction be
  applied?}
\item{\code{calibration_method}:}{integer; which calibration method shall be
  used? Set to \code{0} for simple calibration without drift correction; set
  to \code{1} for simple calibration including linear drift correction; set
  to \code{2} for a double calibration performed at the beginning and the end
  of the measurement sequence including an inherent drift correction.}
\item{\code{use_three_point_calibration}:}{logical; shall three or more
  standards be used in the calibration (\code{TRUE}) or only two
  (\code{FALSE})?}
\item{\code{average_over_inj}:}{the number or range of injections to average
  over in the final processing step. Set to \code{-1} or "all" to use all
  available injections; set to a single integer `n` to use the last `n`
  injections; set to a range `n1:n2` to use the injection range `n1:n2`.}
\item{\code{standards}:}{a list the same length as the number of standards
  used in the measurement data sets; each list is a list with the following
  elements:
  \itemize{
  \item \code{name}: character; the name of the standard as specified by the
    \code{Identifier 1} value in the data set.
  \item \code{o18_True}: numeric; the expected d18O value of the standard.
  \item \code{H2_True}: numeric; the expected dD value of the standard.
  \item \code{use_for_calibration}: logical; shall the standard be used for
    calibration? 
  \item \code{use_for_drift_correction}: logical; shall the standard be used
    to calculate the drift correction? 
  \item \code{use_as_control_standard}: logical; shall the standard be used
    as a quality control standard?
}}
}}
}
\value{
A named list the same length as the input \code{datasets}; each list
  element contains the output of \code{\link{processSingleDataset}}.
}
\description{
This is the main interface for using \code{piccr} as a part of other
programmes to correct and calibrate ("process") a number of similar Picarro
measurement data sets at once. It is is a wrapper for the function
\code{\link{processSingleDataset}} which works on a single data set to do the
actual processing.
}
\details{
This function implements processing of Picarro stable isotope data
including memory correction, drift correction, calibration, calculation of
d-excess values and assessment of information for quality control
monitoring. All processing options are set via parameters in the
\code{config} variable.
}
\seealso{
\code{\link{processSingleDataset}}
}
