% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_processData.R
\name{processData}
\alias{processData}
\title{processData}
\usage{
processData(datasets, config)
}
\arguments{
\item{datasets}{A named list of data.frames. The isotope measurement data
as output by the Picarro device.}

\item{config}{A named list. Required fiels: 
average_over_inj (a number or "all")
$use_three_point_calibration (logical)
$calibration_method (1, 2, or 3)
$use_memory_correction (logical)
$standards (named list of $name (character), $o18_True (numerical), 
            $H2_True (numerical), $use_for_drift_correction (logical), 
            $use_for_calibration (logical)), $use_as_control_standard(logical)}
}
\value{
A list of the same length as \code{datasets}. Each list element is
  again a list of 13 elements with the following structure:
            $name (character vector with the name of the data set (file name))
            $raw (data frame with the raw input isotope data)
            $memoryCorrected (data frame of memory-corrected data)
            $calibrated (data frame of data calibrated without drift correction)
            $calibratedAndDriftCorrected (data frame of calibrated and drift-corrected data)
            $processed (data frame of final processed data)
            $deviationsFromTrue (data frame with the deviations from the true values for all measured standards)
            $rmsdDeviationsFromTrue (named list (elements \code{d18O} and \code{dD}) with the rmsd of deviationsFromTrue)
            $deviationOfControlStandard (named list (elements \code{d18O} and \code{dD}) with the deviation from the true value for the quality control standard)
            $pooledSD (named list (elements \code{d18O} and \code{dD}) of pooled standard deviations for the data set)
            $memoryCoefficients (data frame of memory coefficients)
            $calibrationParams (data frame of calibration regression parameters)
            $driftParams (data frame of drift regression parameters)
}
\description{
The main interface for using piccr as a part of other programs.
}
\details{
Implements the processing of water isotope data using memory correction,
calibration, and drift correction.
Uses the processing options set in the parameter 'config'.
Calculates quality indicators for each dataset such as delta excess and the pooled 
standard deviation.
}
