% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_processFiles.R
\name{processFiles}
\alias{processFiles}
\title{processFiles}
\usage{
processFiles(configFile)
}
\arguments{
\item{configFile}{A character string naming the config file.}
}
\value{
A list. See the documentation for processData(..) for details.
}
\description{
Main interface for the stand-alone use of piccr. Input the path to
a config file. Then process measurement data according to the 
configurations in the config file.
}
\details{
Relies on processData(..) to do the actual processing.
}
