% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_processFiles.R
\name{processFiles}
\alias{processFiles}
\title{Process Picarro CRDS measurement files}
\usage{
processFiles(configFile, writeOutput = TRUE)
}
\arguments{
\item{configFile}{A character string naming the YAML configuration file
that includes the parameter settings to use for the processing; see the
example file provided with \code{piccr} (`system.file("extdata",
"config.yaml", package = "piccr")`) for details.}

\item{writeOutput}{logical; shall processed data and quality control
information be written to disk in the location specified by the
configuration file? Defaults to \code{TRUE}.}
}
\value{
Invisibly, a list the same length as the number of raw Picarro files
  in the input directory; each list element contains the output from
  \code{\link{processSingleDataset}}.
}
\description{
This is the main interface for the stand-alone use of \code{piccr} to process
a number of Picarro data files in a given directory according to the settings
in the configuration file \code{config.yaml}. The function calls
\code{\link{processData}} to do the actual processing.
}
\seealso{
\code{\link{processData}}, \code{\link{processSingleDataset}}
}
