% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_processData.R
\name{processSingleDataset}
\alias{processSingleDataset}
\title{Process a single Picarro CRDS stable isotope data set}
\usage{
processSingleDataset(name, datasets, config)
}
\arguments{
\item{name}{the name of the data set which shall be processed; must match one
of the actual names of \code{datasets}.}

\item{datasets}{a named list of data frames with the raw isotope measurement
data as output by the Picarro device.}

\item{config}{a named list of the following required components:
\describe{
\item{\code{use_memory_correction}.}{logical; shall a memory correction be
  applied?}
\item{\code{calibration_method}:}{integer; which calibration method shall be
  used? Set to \code{0} for simple calibration without drift correction; set
  to \code{1} for simple calibration including linear drift correction; set
  to \code{2} for a double calibration performed at the beginning and the end
  of the measurement sequence including an inherent drift correction.}
\item{\code{use_three_point_calibration}:}{logical; shall three or more
  standards be used in the calibration (\code{TRUE}) or only two
  (\code{FALSE})?}
\item{\code{average_over_inj}:}{the number or range of injections to average
  over in the final processing step. Set to \code{-1} or "all" to use all
  available injections; set to a single integer `n` to use the last `n`
  injections; set to a range `n1:n2` to use the injection range `n1:n2`.}
\item{\code{standards}:}{a list the same length as the number of standards
  used in the measurement data sets; each list is a list with the following
  elements:
  \itemize{
  \item \code{name}: character; the name of the standard as specified by the
    \code{Identifier 1} value in the data set.
  \item \code{o18_True}: numeric; the expected d18O value of the standard.
  \item \code{H2_True}: numeric; the expected dD value of the standard.
  \item \code{use_for_calibration}: logical; shall the standard be used for
    calibration? 
  \item \code{use_for_drift_correction}: logical; shall the standard be used
    to calculate the drift correction? 
  \item \code{use_as_control_standard}: logical; shall the standard be used
    as a quality control standard?
}}
}}
}
\value{
A named list with 13 components:
  \describe{
  \item{\code{name}:}{the file name of the data set.}
  \item{\code{raw}:}{a data frame with the raw measurement data.}
  \item{\code{memoryCorrected}:}{a data frame with the memory-corrected
    measurement data, or \code{NULL} if no memory correction was applied.}
  \item{\code{calibrated}:}{a data frame with the calibrated measurement
    data based either on raw or memory-corrected data.}
  \item{\code{calibratedAndDriftCorrected}:}{a data frame with the calibrated
    and drift-corrected measurement data, or \code{NULL} if no drift
    correction was applied.}
  \item{\code{processed}:}{a data frame with the corrected and calibrated
    measurement data averaged over a specified number of injections.}

  \item{\code{memoryCoefficients}:}{a data frame with the estimated memory
    coefficients.}
  \item{\code{deviationsFromTrue}:}{data frame with the measured (and
    processsed) value (d18O and dH), the expected value, and the deviation of
    the measured from the expected value for each analysed standard.}
  \item{\code{rmsdDeviationsFromTrue}:}{a list with the two elements
    \code{d18O} and \code{dD} with the respective root mean square deviation
    across all deviations in \code{deviationsFromTrue} (with the very first
    standard of the measurement sequence excluded).}
  \item{\code{deviationOfControlStandard}:}{a list with the two elements
    \code{d18O} and \code{dD} of the deviations from the expected value for
    the quality control standard(s) specified in the config file.}
  \item{\code{pooledSD}:}{a list with the two elements \code{d18O}
    and \code{dD} of the pooled standard deviations across all measured
    vials.}
  \item{\code{calibrationParams}:}{calibration parameters; not yet
    implemented.}
  \item{\code{driftParams}:}{estimated drift parameters; not yet
    implemented.}
}
}
\description{
This function implements the processing of a single Picarro stable isotope
data set including memory correction, drift correction, calibration,
calculation of d-excess values and assessment of information for quality
control monitoring. All processing options are set via parameters in the
\code{config} variable.
}
