% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_smallFunctions.R
\name{readFiles}
\alias{readFiles}
\title{Read files}
\usage{
readFiles(config)
}
\arguments{
\item{config}{A named list. Needs to contain at least the#
components 'input_directory' and 'file_extension'.}
}
\value{
A named list of dataframes. The names of the 
        list elements correspond to the file names in the
        input directory.
}
\description{
Read all files from a given input directory that have a 
given file extension. Only works for csv files.
}
\details{
The input directory and the file extension are given in
the config args 'input_directory' and 'file_extension'
respectively.
}
