% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_writeDataToFile.R
\name{removeStandardsFromDataIfRequested}
\alias{removeStandardsFromDataIfRequested}
\title{Remove standards from data frame}
\usage{
removeStandardsFromDataIfRequested(dataset, config)
}
\arguments{
\item{dataset}{a data frame with measurement data of a specific data set.}

\item{config}{a named list containing the component
\code{include_standards_in_output} to signal whether the standard data
shall be removed from \code{dataset}.}
}
\value{
The input \code{dataset} with the standard data removed if signalled
  in \code{config}.
}
\description{
Remove those rows from a given data set which contain the data of the
measured standards.
}
