% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{runCalibrationModel}
\alias{runCalibrationModel}
\title{Run calibration model}
\usage{
runCalibrationModel(trainingData, species = "d18O")
}
\arguments{
\item{trainingData}{An input measurement dataset filtered by the standard
block and the isotope standards which are to be used for the calibration.}

\item{species}{character string with the name of the isotope species for
which the calibration model shall be calculated; valid names are "d18O" for
oxygen isotopes and "dD" for hydrogen isotopes.}
}
\value{
A list with two elements:
\describe{
\item{\code{intercept}:}{numeric; the estimated intercept of the
  calibration.}
\item{\code{slope}:}{numeric; the estimated slope of the calibration.}
}
}
\description{
Run the calibration model of measured standard values against their true
values. Implemented is a simple linear regression.
}
