% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths_linearCalibration.R
\name{selectStandardsForTwoPointCalib}
\alias{selectStandardsForTwoPointCalib}
\title{Select lowest and highest standard}
\usage{
selectStandardsForTwoPointCalib(dataset)
}
\arguments{
\item{dataset}{a data frame with the isotopic data for a set of standards
from a specific block.}
}
\value{
A data frame with all injections from the two selected standards.
}
\description{
From a data set of standards, select the two standards that exhibit the
lowest and highest isotope values.
}
