% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_writeDataToFile.R
\name{writeDatasets}
\alias{writeDatasets}
\title{Write data sets to disc}
\usage{
writeDatasets(folder, datasets, config)
}
\arguments{
\item{folder}{character string with the directory path in which to save the
data.}

\item{datasets}{a list of processed measurement data as output by
\code{\link{processData}} or minimum a list of lists with each sublist
containing a processed data set in the component \code{processed}.}

\item{config}{a named list containing the component
\code{include_standards_in_output}: a logical to signal whether to save
the processed data including the measured standards or not.}
}
\description{
Write the given processed data sets to disc as csv files. This is a wrapper
for \code{\link{writeSingleDataset}} which performs the actual saving for a
single data set.
}
\seealso{
\code{\link{writeSingleDataset}}
}
