% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_writeDataToFile.R
\name{writeDataToFile}
\alias{writeDataToFile}
\title{writeDataToFile}
\usage{
writeDataToFile(datasets, config)
}
\arguments{
\item{datasets}{A nested list, as output by processData(..).
Contains the component $processed.}

\item{config}{A named list. Needs to contain at least the components
'output_directory' and 'include_standards_in_output'.}
}
\value{
No relevant return value
}
\description{
Save processed datasets to disc. Note that the component 
'datasets$processed' contains the dataframes that are
saved.
}
\details{
Uses the following config arguments:
  - 'output_directory'
  - 'include_standards_in_output'
}
